requirements.txt
discord.py==2.3.2
asyncpg==0.29.0
Pillow==10.3.0
python-dotenv==1.0.1
black==24.8.0
isort==5.13.2

